requirements.txt
cowpy==1.0.3
Flask~=1.1.2
flask-cors~=3.0.8
newspaper3k~=0.2.8
